+++ App/Config/DefaultCodes.js
@@ -4,14 +4,18 @@ import { StatusBar, StyleSheet } from 'react-native'
 
 export const commonStyle = StyleSheet.create({
     container: {
-        // flex: 1,
+        flex: 1,
         marginTop: StatusBar.currentHeight,
     },
     formContainer: {
         paddingHorizontal: 10,
+        // flex: 1,
+
     },
     formWrapper: {
         marginTop: 10,
+        // flex: 1,
+
     },
     pageTitle: {
         fontSize: 25,