+++ App/Screens/ChartOfAccountList.js
@@ -95,8 +95,8 @@ const ChartOfAccountList = ({ navigation }) => {
             <ToggleDrawer
                 navigation={navigation}
             />
-            <View style={commonStyle.formContainer}>
-                <View style={commonStyle.formWrapper}>
+            <View style={[commonStyle.formContainer, { flex: 1 }]}>
+                <View style={[commonStyle.formWrapper, { flex: 1 }]}>
                     <View>
                         {editModeActive ? <Text style={commonStyle.pageTitle}>Edit Chart Of Account List</Text> :
                             <Text style={commonStyle.pageTitle}>
@@ -146,94 +146,93 @@ const ChartOfAccountList = ({ navigation }) => {
                         </View>
                     </View>
 
-                    {editModeActive ?
-                        <View style={{ marginTop: 10, alignItems: "center" }}>
+                    {
+                        editModeActive ?
+                            <View style={{ marginTop: 10, alignItems: "center" }}>
 
-                            <TextInput
-                                style={commonStyle.textInput}
-                                defaultValue={defaultValue.oldName}
-                                value={textInput}
-                                onChangeText={text => setTextInput(text)}
-                            />
-                            <View style={commonStyle.checkBox}>
-                                <CheckBox
-                                    disabled={false}
-                                    tintColors={{ true: '#F54748', false: 'black' }}
-                                    value={toggleCheckBox}
-                                    onValueChange={(newValue) => setToggleCheckBox(newValue)}
+                                <TextInput
+                                    style={commonStyle.textInput}
+                                    defaultValue={defaultValue.oldName}
+                                    value={textInput}
+                                    onChangeText={text => setTextInput(text)}
                                 />
-                                <Text style={{ fontSize: 18, }}>Active</Text>
-                            </View>
-                            <TouchableOpacity
-                                style={[commonStyle.submitORsaveBtn, { width: "80%", backgroundColor: "#00A19D" }]}
-                                onPress={() => cancelUpdate()}
-                            >
-                                <Text style={{ color: "#fff", fontSize: 18 }}>
-                                    Cancel
-                                </Text>
-                            </TouchableOpacity>
-                            <TouchableOpacity
-                                style={[commonStyle.submitORsaveBtn, { width: "80%" }]}
-                                onPress={() => updateCatagory()}>
-                                <Text style={{ color: "#fff", fontSize: 18 }}>
-                                    Update
-                                </Text>
-                            </TouchableOpacity>
-
-                        </View>
-                        : <View style={{
-                            // maxHeight: 500,
-                            // width: "95%",
-                            // marginHorizontal: 10,
-                            // flex: 1,
-                            // flexGrow: 1
-                        }}>
-                            {subCatagories ?
-                                <View style={{
-                                    width: "95%",
-                                    marginHorizontal: 10,
-                                }}>
-                                    <Text style={{
-                                        textAlign: "center",
-                                        marginTop: 10,
-                                        fontSize: 17
-                                    }}>Available sub catagories</Text>
-                                    <FlatList
-                                        style={{
-                                            marginTop: 10,
-                                            paddingVertical: 10,
-                                            paddingHorizontal: 10,
-                                            borderWidth: 3,
-                                            borderRadius: 10,
-                                            borderColor: "#DDDDDD",
-                                        }}
-                                        data={subCatagories}
-                                        keyExtractor={(item, index) => index.toString()}
-                                        renderItem={({ item }) => (
-                                            <View style={{
-                                                flexDirection: "row",
-                                                justifyContent: "space-between",
-                                                marginBottom: 5,
-                                                marginTop: 10,
-                                            }}>
-                                                <View>
-                                                    <Text style={{
-                                                        fontSize: 20,
-                                                    }}>{item.name}</Text>
-                                                </View>
-
-                                                <TouchableOpacity style={commonStyle.smallButton}
-                                                    onPress={() => editTheSubCatagory(item.id, item.name, item.parentId)}
-                                                >
-                                                    <Text style={{ color: "#fff" }}>Edit</Text>
-                                                </TouchableOpacity>
-                                            </View>
-                                        )}
+                                <View style={commonStyle.checkBox}>
+                                    <CheckBox
+                                        disabled={false}
+                                        tintColors={{ true: '#F54748', false: 'black' }}
+                                        value={toggleCheckBox}
+                                        onValueChange={(newValue) => setToggleCheckBox(newValue)}
                                     />
-                                </View> : null}
-                        </View>}
+                                    <Text style={{ fontSize: 18, }}>Active</Text>
+                                </View>
+                                <TouchableOpacity
+                                    style={[commonStyle.submitORsaveBtn, { width: "80%", backgroundColor: "#00A19D" }]}
+                                    onPress={() => cancelUpdate()}
+                                >
+                                    <Text style={{ color: "#fff", fontSize: 18 }}>
+                                        Cancel
+                                    </Text>
+                                </TouchableOpacity>
+                                <TouchableOpacity
+                                    style={[commonStyle.submitORsaveBtn, { width: "80%" }]}
+                                    onPress={() => updateCatagory()}>
+                                    <Text style={{ color: "#fff", fontSize: 18 }}>
+                                        Update
+                                    </Text>
+                                </TouchableOpacity>
+
+                            </View>
+                            : null
+                    }
+
+                    {subCatagories ?
+                        <View style={{
+                            width: "95%",
+                            marginHorizontal: 10,
+                            flex: 1,
+                            marginBottom: 10,
+                        }}>
+                            <Text style={{
+                                textAlign: "center",
+                                marginTop: 10,
+                                fontSize: 17
+                            }}>Available sub catagories</Text>
+                            <View style={{
+                                flex: 1,
+                            }}>
+                                <FlatList
+                                    style={{
+                                        marginTop: 10,
+                                        paddingHorizontal: 10,
+                                    }}
+                                    data={subCatagories}
+                                    keyExtractor={(item, index) => index.toString()}
+                                    renderItem={({ item }) => (
+                                        <View style={{
+                                            flexDirection: "row",
+                                            justifyContent: "space-between",
+                                            marginBottom: 10,
+                                            marginTop: 10,
+                                        }}>
+                                            <View>
+                                                <Text style={{
+                                                    fontSize: 20,
+                                                }}>{item.name}</Text>
+                                            </View>
+
+                                            <TouchableOpacity style={commonStyle.smallButton}
+                                                onPress={() => editTheSubCatagory(item.id, item.name, item.parentId)}
+                                            >
+                                                <Text style={{ color: "#fff" }}>Edit</Text>
+                                            </TouchableOpacity>
+                                        </View>
+                                    )}
+                                />
+                            </View>
+                        </View> : null}
                 </View>
             </View>
+
         </View >
     )
 }