+++ App/Screens/EditChartOfAccount.js
@@ -1,22 +1,24 @@
+
 import CheckBox from '@react-native-community/checkbox';
 import React from 'react';
-import { Text, TextInput, TouchableOpacity, View } from 'react-native';
+import { TextInput, TouchableOpacity, View } from 'react-native';
 import { commonStyle } from './../Config/DefaultCodes';
 
-const EditChartOfAccount = ({ defaultValue, setTextInput,
-    textInput, toggleCheckBox, setToggleCheckBox,
-    updateCatagory }) => {
-    return (
+const EditChartOfAccount = ({ defaultValue, textInput, setTextInput,
+    toggleCheckBox, setToggleCheckBox
+    , cancelUpdate, updateCatagory }) => {
 
-        <View style={{ marginTop: 10, alignItems: "center" }
-        } >
+
+        
+    return (
+        <View style={{ marginTop: 10, alignItems: "center" }}>
+
             <TextInput
                 style={commonStyle.textInput}
-                defaultValue={defaultValue?.oldName}
+                defaultValue={defaultValue.oldName}
                 value={textInput}
                 onChangeText={text => setTextInput(text)}
             />
-
             <View style={commonStyle.checkBox}>
                 <CheckBox
                     disabled={false}
@@ -27,14 +29,22 @@ const EditChartOfAccount = ({ defaultValue, setTextInput,
                 <Text style={{ fontSize: 18, }}>Active</Text>
             </View>
             <TouchableOpacity
+                style={[commonStyle.submitORsaveBtn, { width: "80%", backgroundColor: "#00A19D" }]}
+                onPress={() => cancelUpdate()}
+            >
+                <Text style={{ color: "#fff", fontSize: 18 }}>
+                    Cancel
+                </Text>
+            </TouchableOpacity>
+            <TouchableOpacity
                 style={[commonStyle.submitORsaveBtn, { width: "80%" }]}
                 onPress={() => updateCatagory()}>
                 <Text style={{ color: "#fff", fontSize: 18 }}>
                     Update
                 </Text>
             </TouchableOpacity>
-        </View >
 
+        </View>
     )
 }