+++ App/Screens/Porfile.js
@@ -1,16 +1,15 @@
 import React, { useContext } from 'react';
-import { StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { Image, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { AppContext } from '../Context/AppContext';
 import { AuthContext } from '../Context/AuthContext';
+import { commonStyle } from './../Config/DefaultCodes';
 import ToggleDrawer from './ToggleDrawer';
+
 
 const Porfile = ({ navigation }) => {
     const { userLogInInfo } = useContext(AppContext)
     const { logout } = useContext(AuthContext)
-
-    // const logoutUser = () => {
-    //     setLoggedInUser(false)
-    // }
+    console.log(userLogInInfo)
 
     return (
         <View style={styles.conatiner}>
@@ -19,19 +18,25 @@ const Porfile = ({ navigation }) => {
                 backgroundColor="#fff"
                 translucent={true}
             />
-             <ToggleDrawer
+            <ToggleDrawer
                 navigation={navigation}
             />
-            <View style={styles.details}>
-                <View>
-                    <Text>Email: {userLogInInfo.data.email}</Text>
-                    <Text>Name: {userLogInInfo.data.fullName}</Text>
-                    <Text>Phone: {userLogInInfo.data.phone}</Text>
+            <View style={{ alignItems: "center" }}>
+                <Image
+                    style={{ width: 100, height: 100, marginTop: 50, }}
+                    source={require('../assets/icons/userProfile.png')}
+                />
+                <View style={{ alignItems: "center", marginTop: 50, }}>
+                    <Text style={{ fontSize: 20 }}>{userLogInInfo.data.email}</Text>
+                    <Text style={{ fontSize: 20 }}>{userLogInInfo.data.fullName}</Text>
+                    <Text style={{ fontSize: 20 }}>{userLogInInfo.data.phone}</Text>
                 </View>
-                <TouchableOpacity onPress={() => logout()} style={styles.logoutBtn}>
-                    <Text style={styles.logoutBtnText}>
-                        Logout
-                    </Text>
+            </View>
+            <View style={{ marginHorizontal: 20, }}>
+                <TouchableOpacity
+                    onPress={() => logout()}
+                    style={[commonStyle.submitORsaveBtn, { backgroundColor: "#3DB2FF" }]}>
+                    <Text style={{ fontSize: 20, color: "#fff" }}>Logout</Text>
                 </TouchableOpacity>
             </View>
 
@@ -45,26 +50,6 @@ const styles = StyleSheet.create({
     conatiner: {
         flex: 1,
         marginTop: StatusBar.currentHeight,
-        alignItems: "center",
-        justifyContent: "center"
     },
-    details: {
-        padding: 10,
-        width: "60%",
-        borderRadius: 10,
-    },
-    logoutBtn: {
-        alignItems: "center",
-        justifyContent: "center",
-        marginTop: 10,
-        width: "100%",
-        height: 40,
-        borderRadius: 10,
-        backgroundColor: "#345B63",
-    },
-    logoutBtnText: {
-        fontSize: 18,
-        color: "#fff",
-        textAlign: "center"
-    },
+
 })